process : LC026

+++ src/bot/utils/helpers/delete-prev-message.ts
@@ -1,9 +1,9 @@
 import { MyContext } from '@/bot/types';
 
 export async function deletePrevMessage(ctx: MyContext) {
-  if (ctx['message']) {
+  if (ctx['message'] && ctx['message']['chat']['id']) {
     await ctx.api.deleteMessage(ctx.message.chat.id, ctx.message.message_id);
-  } else if (ctx['callbackQuery']) {
+  } else if (ctx['callbackQuery'] && ctx['callbackQuery']['message']['chat']['id']) {
     await ctx.api.deleteMessage(
       ctx.callbackQuery.message.chat.id,
       ctx.callbackQuery.message.message_id,
